requirements.txt
tqdm
lmdb
pillow
cupy
numpy
einops
opencv-python
matplotlib
prefetch-generator
glob
joblib
tensorboard